requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
django-bootstrap5==22.2
django-scheduler==0.10.1
django-tempus-dominus==5.1.2.17
icalendar==5.0.4
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
